requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
blinker==1.6.3
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cryptography==41.0.4
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flower==2.0.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
gunicorn==21.2.0
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
locust==2.17.0
MarkupSafe==2.1.3
model-bakery==1.16.0
msgpack==1.0.7
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
rpds-py==0.10.6
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.1
